Templates/scripts/jdl_generator/requirements.txt
lark-parser>=0.5.6
Jinja2>=2.8